+++ resources/views/task/show.blade.php
@@ -25,7 +25,9 @@
         </dd>
     </dl>
     @auth
-        <a class="btn btn-primary btn-lg" href="{{ route('tasks.destroy', $task) }}" data-confirm="are you sure?" data-method="delete" rel="nofollow"  role="button">{{ __('Remove') }}</a>
+        @if($task->creator->id == Auth::id())
+            <a class="btn btn-primary btn-lg" href="{{ route('tasks.destroy', $task) }}" data-confirm="are you sure?" data-method="delete" rel="nofollow"  role="button">{{ __('Remove') }}</a>
+        @endif
         <a class="btn btn-primary btn-lg" href="{{ route('tasks.edit', $task) }}" role="button" >{{ __('Edit') }}</a>
     @endauth
 </div>
